requirements.txt
requests>=2.31.0
httpx>=0.25.0
python-dotenv>=1.0.0
openai>=1.12.0
tiktoken>=0.5.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
streamlit>=1.31.0
plotly>=5.18.0
altair>=5.2.0
tqdm>=4.66.0
loguru>=0.7.0
tenacity>=8.2.0
pytest>=7.4.0
black>=24.0.0
